チーム別キャラクターの実装と自分のターン中にブロックを送信するよう変更
