--- Assets/Resources/Player.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 679875af552a4ea42b92343637b8fe4f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
