+++ Assets/Resources/PlayerA.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a8795d3bef5270e42b2340d7b6bdca30
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
