+++ Assets/Scripts/PhotonLogin.cs
@@ -1,32 +1,12 @@
 using Photon.Pun;
 using Photon.Realtime;
 using UnityEngine;
-using System.Runtime.InteropServices;
-using TMPro;
 
 public class PhotonLogin : MonoBehaviourPunCallbacks
 {
-    [DllImport("__Internal")]
-    private static extern void doCode();
-
-    [DllImport("__Internal")]
-    private static extern void setTargetObject(string str);
-
-    [DllImport("__Internal")]
-    private static extern string getBlockFromWorkspace();
-
-    [DllImport("__Internal")]
-    private static extern void setFriendBlock(string block);
-
-    [DllImport("__Internal")]
-    private static extern void setRivalBlock(string block);
-
     public GameObject obstacle;
     [SerializeField] private GameObject TeamSelectPanel;
-    [SerializeField] private TextMeshProUGUI MyTeamLabel;
-    [SerializeField] private TextMeshProUGUI RivalTeamLabel;
     private bool PlayingFlag;
-    private float elapsedTime;
 
     public static PhotonLogin instance;
 
@@ -37,7 +17,6 @@ public class PhotonLogin : MonoBehaviourPunCallbacks
             instance = this;
         }
         PlayingFlag = false;
-        elapsedTime = 0f;
     }
 
     private void Start()
@@ -60,75 +39,15 @@ public class PhotonLogin : MonoBehaviourPunCallbacks
     {
         if (PhotonNetwork.IsMasterClient)
         {
-            var position = new Vector3(0, 0, 0);
-            PhotonNetwork.InstantiateRoomObject("Player", position, Quaternion.identity);
+            var positionA = new Vector3(0, 0, 0);
+            var positionB = new Vector3(5, 0, 5);
+            PhotonNetwork.InstantiateRoomObject("PlayerA", positionA, Quaternion.identity);
+            PhotonNetwork.InstantiateRoomObject("PlayerB", positionB, Quaternion.identity);
 
             PhotonNetwork.InstantiateRoomObject(obstacle.name, new Vector3(-1f, 0, 1f), Quaternion.identity);
             PhotonNetwork.InstantiateRoomObject(obstacle.name, new Vector3(1f, 0, 1f), Quaternion.identity);
             PhotonNetwork.InstantiateRoomObject(obstacle.name, new Vector3(1f, 0, 2f), Quaternion.identity);
         }
-    }
-
-    public void DoCode()
-    {
-        Debug.Log("実行");
-        GameObject obj =  GameObject.FindGameObjectWithTag("Player");
-        PhotonView photonView = obj.GetComponent<PhotonView>();
-        if (!photonView.IsMine)
-        {
-            photonView.RequestOwnership();
-        }
-#if !UNITY_EDITOR && UNITY_WEBGL
-        setTargetObject(obj.name);
-        doCode();
-#endif
-    }
-
-    public void SendMyBlock()
-    {
-#if !UNITY_EDITOR && UNITY_WEBGL
-        photonView.RPC(nameof(SetOthersBlock), RpcTarget.Others, getBlockFromWorkspace());
-#endif
-    }
-
-    [PunRPC]
-    public void SetOthersBlock(string block, PhotonMessageInfo info)
-    {
-        if (!TurnManager.instance.CheckMyTurn(1))
-        {
-            Debug.Log("setOthersBlock実行");
-            if (MyTeamLabel.text == info.Sender.GetTeam())
-            {
-#if !UNITY_EDITOR && UNITY_WEBGL
-                setFriendBlock(block);
-#endif
-            }
-            else
-            {
-#if !UNITY_EDITOR && UNITY_WEBGL
-                setRivalBlock(block);
-#endif
-            }
-        }
-    }
-
-    private void Update()
-    {
-        if (PhotonNetwork.CurrentRoom != null)//ルーム内である
-        {
-            GameObject obj = GameObject.FindGameObjectWithTag("Player");
-            PhotonView photonView = obj.GetComponent<PhotonView>();
-            if (photonView.IsMine)
-            {
-                // 0.5秒毎にテキストを更新する
-                elapsedTime += Time.deltaTime;
-                if (elapsedTime > 0.5f)
-                {
-                    elapsedTime = 0f;
-                    SendMyBlock();
-                }
-            }
-        }   
     }
 
     public void FocusCanvas(string p_focus)
@@ -145,76 +64,14 @@ public class PhotonLogin : MonoBehaviourPunCallbacks
 #endif
     }
 
-    public void GameStart()
-    {
-        // ルーム内のメンバー全員が準備完了状態のときのみ押せるようにする
-        int Anum = 0, Bnum = 0;
-        foreach (var player in PhotonNetwork.PlayerList)
-        {
-            if(player.GetTeam() == "A")
-            {
-                Anum++;
-                player.SetOrder(Anum);
-            }
-            else
-            {
-                Bnum++;
-                player.SetOrder(Bnum);
-            }
-        }
-        if(Anum == 0 || Bnum == 0)
-        {
-            Debug.Log("人数が足りていません");
-        }
-        else
-        {
-            // 乱数で先攻チームを決める
-            if (Random.Range(0, 2) == 0)
-            {
-                PhotonNetwork.CurrentRoom.SetFirst("A");
-            }
-            else
-            {
-                PhotonNetwork.CurrentRoom.SetFirst("B");
-            }
-            PhotonNetwork.CurrentRoom.SetANum(Anum);
-            PhotonNetwork.CurrentRoom.SetBNum(Bnum);
-            photonView.RPC(nameof(RPCGameStart), RpcTarget.AllViaServer);
-        }
-    }
-
-    [PunRPC]
-    public void RPCGameStart()
-    {
-        Debug.Log("ゲーム開始");
-        TeamSelectPanel.transform.localScale = Vector3.zero;
-        PlayingFlag = true;
-        if (PhotonNetwork.IsMasterClient)
-        {
-            TurnManager.instance.StartTurn();
-        }
-    }
-
-    public void SetTeamLabel(string myTeam)
-    {
-        MyTeamLabel.text = myTeam;
-        if (MyTeamLabel.text == "A")
-        {
-            RivalTeamLabel.text = "B";
-        }
-        else
-        {
-            RivalTeamLabel.text = "A";
-        }
-    }
-
     public bool GetPlayingFlag()
     {
         return PlayingFlag;
     }
 
-    public string GetMyTeamLabel()
+    public void GameInit()
     {
-        return MyTeamLabel.text;
+        PlayingFlag = true;
+        TeamSelectPanel.transform.localScale = Vector3.zero;
     }
 }
